--- NotesTakerApp/NotesTakerApp.Presentation/ViewModel/UserViewModel.cs
@@ -1,15 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace NotesTakerApp.Presentation.ViewModel
-{
-    public class UserViewModel
-    {
-    public string UserName { get; set; }
-    public string Email { get; set; }
-    public string Password { get; set; }
-    public string Roles { get; set; }
-    }
-}
