+++ NotesTakerApp/NotesTakerApp.Presentation/Views/Home/Index.cshtml
@@ -1,8 +1,16 @@
 ﻿@{
-    ViewData["Title"] = "Home Page";
+    ViewData["Title"] = "Welcome to Notes Taker";
 }
 
-<div class="text-center">
-    <h1 class="display-4">Welcome</h1>
-    <p>Learn about <a href="https://learn.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
+<div class="text-center mt-5">
+    <h1 class="display-4">Welcome to Notes Taker 📝</h1>
+    <p class="lead mt-3">
+        Your personal space to capture, organize, and manage all your notes effortlessly.
+    </p>
+
+    <div class="mt-4">
+        <a class="btn btn-primary btn-lg me-2" href="/Identity/Register">Get Started</a>
+        <a class="btn btn-outline-secondary btn-lg" href="/Identity/Login">Login</a>
+    </div>
 </div>
+
