+++ NotesTakerApp/NotesTakerApp.Presentation/Views/Identity/Login.cshtml
@@ -1,5 +1,4 @@
-
-@using NotesTakerApp.Core.ViewModels
+@using NotesTakerApp.Core.Models
 @model UserViewModel
 
 <form method="post" asp-controller="Identity" asp-action="Login">
