+++ NotesTakerApp/NotesTakerApp.Presentation/Views/Identity/Register.cshtml
@@ -1,24 +1,24 @@
-@using NotesTakerApp.Presentation.ViewModel
+@using NotesTakerApp.Core.Models
 @model UserViewModel
 
 
-<form method="post" asp-controller="Account" asp-action="Register">
+<form method="post" asp-controller="Identity" asp-action="Register">
     <div asp-validation-summary="All"></div>
     <div class="mb-3">
-        <label asp-for="UserName" class="form-label">Username</label>
-        <input asp-for="UserName" class="form-control">
-        <span class="text-danger" asp-validation-for="UserName"></span>
+        <label asp-for="NewUser.UserName" class="form-label">Username</label>
+        <input asp-for="NewUser.UserName" class="form-control">
+        <span class="text-danger" asp-validation-for="NewUser.UserName"></span>
     </div>
     
     <div class="mb-3">
-        <label asp-for="Email" class="form-label">Email</label>
-        <input asp-for="Email" class="form-control">
-        <span class="text-danger" asp-validation-for="Email"></span>
+        <label asp-for="NewUser.Email" class="form-label">Email</label>
+        <input asp-for="NewUser.Email" class="form-control">
+        <span class="text-danger" asp-validation-for="NewUser.Email"></span>
     </div>
     <div class="mb-3">
-        <label asp-for="Password" class="form-label">Password</label>
-        <input type="password" asp-for="Password" class="form-control">
-        <span class="text-danger" asp-validation-for="Password"></span>
+        <label asp-for="NewUser.PasswordHash" class="form-label">Password</label>
+        <input type="password" asp-for="NewUser.PasswordHash" class="form-control">
+        <span class="text-danger" asp-validation-for="NewUser.PasswordHash"></span>
     </div>
 
     <button type="submit" class="btn btn-primary">Submit</button>
