+++ NotesTakerApp/NotesTakerApp.Presentation/Views/Shared/_Layout.cshtml
@@ -55,12 +55,6 @@
         <main role="main" class="pb-3">
             @RenderBody()
         </main>
-
-        <footer class="border-top footer text-muted mt-5">
-            <div class="container">
-                &copy; 2025 - NotesTakerApp.Presentation - <a asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
-            </div>
-        </footer>
     </div>
 
     <!-- Scripts -->
