+++ NotesTakerApp/NotesTakerApp.Presentation/Views/User/Index.cshtml
@@ -1,4 +1,4 @@
-@model NotesTakerApp.Core.ViewModels.UserViewModel
+@model NotesTakerApp.Core.Models.UserViewModel
 
 @{
     ViewData["Title"] = "User Management";
